Producer Consumer for logging, Log to file plugin, Log to Remote Api plugin.

+++ SystemMonitor.Core/Models/DiskUsageDto.cs
@@ -1,8 +1,11 @@
 namespace SystemMonitor.Core.Models;
 
+/// <summary>
+/// Data Object for disk UUsage
+/// </summary>
 public class DiskUsageDto
 {
     public required string DriveName { get; set; }
-    public double DiskUsed { get; set; }
-    public double DiskTotal { get; set; }
+    public long DiskUsed { get; set; }
+    public long DiskTotal { get; set; }
 }
